chore: prettify code and changed visualEditorWrapper to visualEditorContainer
