+++ src/liveEditor/generators/generateAddInstanceButtons.tsx
@@ -9,18 +9,18 @@ import AddInstanceButtonComponent from "../components/addInstanceButton";
 export function generateAddInstanceButton(
     onClickCallback: (event: MouseEvent) => void
 ): HTMLButtonElement {
-    const visualEditorWrapper = document.querySelector(".visual-editor__container");
+    const visualEditorContainer = document.querySelector(".visual-editor__container");
     const wrapper = document.createDocumentFragment();
     let button = document.querySelector(
         ".visual-editor__add-button"
         ) as HTMLButtonElement;
 
-    if (!visualEditorWrapper?.contains(button)) {
+    if (!visualEditorContainer?.contains(button)) {
         render(
             <AddInstanceButtonComponent onClickCallback={onClickCallback} />,
             wrapper
         );
-        visualEditorWrapper?.appendChild(wrapper);
+        visualEditorContainer?.appendChild(wrapper);
     }
 
     button = document.querySelector(
@@ -31,26 +31,26 @@ export function generateAddInstanceButton(
 }
 
 export function getAddInstanceButtons(
-    visualEditorWrapper: HTMLDivElement,
+    visualEditorContainer: HTMLDivElement,
     getAllButtons: true
 ): HTMLButtonElement[];
 
 export function getAddInstanceButtons(
-    visualEditorWrapper: HTMLDivElement,
+    visualEditorContainer: HTMLDivElement,
     getAllButtons?: false
 ): [HTMLButtonElement, HTMLButtonElement] | null;
 
 /**
  * Returns an array of HTMLButtonElement instances that can be used to add new instances to the visual editor.
- * @param visualEditorWrapper - The HTMLDivElement that contains the visual editor.
+ * @param visualEditorContainer - The HTMLDivElement that contains the visual editor.
  * @param getAllButtons - If true, returns all add instance buttons. If false, returns only the previous and next buttons.
  * @returns An array of HTMLButtonElement instances or null if there are less than 2 buttons.
  */
 export function getAddInstanceButtons(
-    visualEditorWrapper: HTMLDivElement,
+    visualEditorContainer: HTMLDivElement,
     getAllButtons = false
 ): HTMLButtonElement[] | [HTMLButtonElement, HTMLButtonElement] | null {
-    const buttons = visualEditorWrapper.getElementsByClassName(
+    const buttons = visualEditorContainer.getElementsByClassName(
         "visual-editor__add-button"
     );
