+++ src/liveEditor/generators/generateCustomCursor.ts
@@ -1,4 +1,4 @@
-import { FieldDataType } from "./types/index.types";
+import { FieldDataType } from "../utils/types/index.types";
 
 const modular_block = `<svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg"><path fill-rule="evenodd" clip-rule="evenodd" d="M2.25 3.5625C2.25 2.83763 2.83763 2.25 3.5625 2.25H20.441C21.1659 2.25 21.7535 2.83763 21.7535 3.5625V20.441C21.7535 21.1659 21.1659 21.7535 20.441 21.7535H3.5625C2.83763 21.7535 2.25 21.1659 2.25 20.441V3.5625ZM3.375 12.5643V20.441C3.375 20.5446 3.45895 20.6285 3.5625 20.6285H11.4393V12.5643L3.375 12.5643ZM11.4393 11.4393L3.375 11.4393V3.5625C3.375 3.45895 3.45895 3.375 3.5625 3.375H11.4393V11.4393ZM12.5643 12.5643V20.6285H20.441C20.5446 20.6285 20.6285 20.5446 20.6285 20.441V12.5643L12.5643 12.5643ZM20.6285 11.4393L12.5643 11.4393V3.375H20.441C20.5446 3.375 20.6285 3.45895 20.6285 3.5625V11.4393Z" fill="white"/></svg>`;
 const url = `<svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg"><path fill-rule="evenodd" clip-rule="evenodd" d="M15.5747 8.4136C15.7945 8.63311 15.7948 8.98927 15.5753 9.2091L9.21902 15.5747C8.99952 15.7946 8.64336 15.7948 8.42353 15.5753C8.2037 15.3558 8.20344 14.9996 8.42294 14.7798L14.7792 8.41419C14.9987 8.19436 15.3549 8.1941 15.5747 8.4136Z" fill="white"/><path fill-rule="evenodd" clip-rule="evenodd" d="M7.6251 10.0076C7.84477 10.2273 7.84477 10.5835 7.6251 10.8031L4.97197 13.4563C4.23333 14.1949 3.81836 15.1967 3.81836 16.2413C3.81836 17.2859 4.23333 18.2878 4.97197 19.0264C5.71062 19.765 6.71243 20.18 7.75704 20.18C8.27427 20.18 8.78644 20.0781 9.2643 19.8802C9.74216 19.6823 10.1764 19.3921 10.5421 19.0264L13.1952 16.3733C13.4149 16.1536 13.7711 16.1536 13.9907 16.3733C14.2104 16.5929 14.2104 16.9491 13.9907 17.1688L11.3376 19.8219C10.8674 20.2921 10.3092 20.6651 9.69482 20.9196C9.08047 21.174 8.42201 21.305 7.75704 21.305C6.41406 21.305 5.1261 20.7715 4.17648 19.8219C3.22685 18.8723 2.69336 17.5843 2.69336 16.2413C2.69336 14.8984 3.22685 13.6104 4.17648 12.6608L6.8296 10.0076C7.04927 9.78798 7.40543 9.78798 7.6251 10.0076Z" fill="white"/><path fill-rule="evenodd" clip-rule="evenodd" d="M12.6608 4.17648C13.6104 3.22685 14.8984 2.69336 16.2413 2.69336C17.5843 2.69336 18.8723 3.22685 19.8219 4.17648C20.7715 5.1261 21.305 6.41406 21.305 7.75704C21.305 9.10001 20.7715 10.388 19.8219 11.3376L17.1688 13.9907C16.9491 14.2104 16.5929 14.2104 16.3733 13.9907C16.1536 13.7711 16.1536 13.4149 16.3733 13.1952L19.0264 10.5421C19.765 9.80345 20.18 8.80164 20.18 7.75704C20.18 6.71243 19.765 5.71062 19.0264 4.97197C18.2878 4.23333 17.2859 3.81836 16.2413 3.81836C15.1967 3.81836 14.1949 4.23333 13.4563 4.97197L10.8031 7.6251C10.5835 7.84477 10.2273 7.84477 10.0076 7.6251C9.78798 7.40543 9.78798 7.04927 10.0076 6.8296L12.6608 4.17648Z" fill="white"/></svg>`;
