+++ src/liveEditor/globals/index.ts
@@ -2,7 +2,7 @@ interface VisualEditorGlobalUtilsImpl {
     // customCursor: HTMLDivElement | null;
     // overlayWrapper: HTMLDivElement | null;
     previousSelectedEditableDOM: HTMLElement | Element | null;
-    // visualEditorWrapper: HTMLDivElement | null;
+    // visualEditorContainer: HTMLDivElement | null;
     previousHoveredTargetDOM: Element | null;
     // focusedToolbar: HTMLDivElement | null;
 }
@@ -11,7 +11,7 @@ const VisualEditorGlobalUtils: VisualEditorGlobalUtilsImpl = {
     // customCursor: null,
     // overlayWrapper: null,
     previousSelectedEditableDOM: null,
-    // visualEditorWrapper: null,
+    // visualEditorContainer: null,
     previousHoveredTargetDOM: null,
     // focusedToolbar: null
 };
