+++ src/liveEditor/listeners/mouseOver.ts
@@ -1,4 +1,4 @@
-import { addCslpOutline } from "../../../cslp";
+import { addCslpOutline } from "../../cslp";
 
 function handleMouseOver(event: MouseEvent) {
     addCslpOutline(event);
